helpers in data_structures
